+++ airflow/api_connexion/endpoints/backfill_endpoint.py
@@ -23,7 +23,7 @@ from typing import TYPE_CHECKING, cast
 
 import pendulum
 from pendulum import DateTime
-from sqlalchemy import func, select
+from sqlalchemy import select
 
 from airflow.api_connexion import security
 from airflow.api_connexion.exceptions import Conflict, NotFound
@@ -32,13 +32,9 @@ from airflow.api_connexion.schemas.backfill_schema import (
     backfill_collection_schema,
     backfill_schema,
 )
-from airflow.exceptions import AirflowException
-from airflow.models.backfill import Backfill, BackfillDagRun
-from airflow.models.serialized_dag import SerializedDagModel
+from airflow.models.backfill import AlreadyRunningBackfill, Backfill, _create_backfill
 from airflow.utils import timezone
 from airflow.utils.session import NEW_SESSION, provide_session
-from airflow.utils.state import DagRunState
-from airflow.utils.types import DagRunTriggeredByType, DagRunType
 from airflow.www.decorators import action_logging
 
 if TYPE_CHECKING:
@@ -66,87 +62,6 @@ def backfill_to_dag(func):
         return func(dag_id=backfill.dag_id, backfill_id=backfill_id, session=session, **kwargs)
 
     return wrapper
-
-
-class AlreadyRunningBackfill(AirflowException):
-    """Raised when attempting to create backfill and one already active."""
-
-
-@provide_session
-def _create_backfill(
-    *,
-    dag_id: str,
-    from_date: DateTime,
-    to_date: DateTime,
-    max_active_runs: int,
-    reverse: bool,
-    dag_run_conf: dict | None,
-    session: Session = NEW_SESSION,
-) -> Backfill | None:
-    serdag = session.get(SerializedDagModel, dag_id)
-    if not serdag:
-        raise NotFound(f"Could not find dag {dag_id}")
-
-    num_active = session.scalar(
-        select(func.count()).where(Backfill.dag_id == dag_id, Backfill.completed_at.is_(None))
-    )
-    if num_active > 0:
-        raise AlreadyRunningBackfill(
-            f"Another backfill is running for dag {dag_id}. "
-            f"There can be only one running backfill per dag."
-        )
-
-    br = Backfill(
-        dag_id=dag_id,
-        from_date=from_date,
-        to_date=to_date,
-        max_active_runs=max_active_runs,
-        dag_run_conf=dag_run_conf,
-    )
-    session.add(br)
-    session.commit()
-
-    dag = serdag.dag
-    depends_on_past = any(x.depends_on_past for x in dag.tasks)
-    if depends_on_past:
-        if reverse is True:
-            raise ValueError(
-                "Backfill cannot be run in reverse when the dag has tasks where depends_on_past=True"
-            )
-
-    backfill_sort_ordinal = 0
-    dagrun_info_list = dag.iter_dagrun_infos_between(from_date, to_date)
-    if reverse:
-        dagrun_info_list = reversed([x for x in dag.iter_dagrun_infos_between(from_date, to_date)])
-    for info in dagrun_info_list:
-        backfill_sort_ordinal += 1
-        log.info("creating backfill dag run %s dag_id=%s backfill_id=%s, info=", dag.dag_id, br.id, info)
-        dr = None
-        try:
-            dr = dag.create_dagrun(
-                triggered_by=DagRunTriggeredByType.BACKFILL,
-                execution_date=info.logical_date,
-                data_interval=info.data_interval,
-                start_date=timezone.utcnow(),
-                state=DagRunState.QUEUED,
-                external_trigger=False,
-                conf=br.dag_run_conf,
-                run_type=DagRunType.BACKFILL_JOB,
-                creating_job_id=None,
-                session=session,
-            )
-        except Exception:
-            dag.log.exception("something failed")
-            session.rollback()
-        session.add(
-            BackfillDagRun(
-                backfill_id=br.id,
-                dag_run_id=dr.id if dr else None,  # this means we failed to create the dag run
-                sort_ordinal=backfill_sort_ordinal,
-            )
-        )
-        session.commit()
-    return br
 
 
 @security.requires_access_dag("GET")
